requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
frozenlist==1.4.1
idna==3.7
lxml==5.2.2
magic-filter==1.0.12
multidict==6.0.5
pillow==10.4.0
pip==24.1.2
pydantic-core==2.20.1
pydantic==2.8.2
python-pptx==1.0.2
setuptools==65.5.1
typing-extensions==4.12.2
wheel==0.38.4
XlsxWriter==3.2.0
yarl==1.9.4
